requirements.txt
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.1
django-admin-autocomplete-filter==0.7.1
django-debug-toolbar==3.7.0
django-filter==22.1
django-nose==1.4.7
django-rest-swagger==2.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
entrypoints==0.4
executing==0.8.3
factory-boy==3.2.0
Faker==8.12.1
fastjsonschema==2.15.3
fonttools==4.33.3
gunicorn==20.1.0
idna==3.2
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.1
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.2.2
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
MarkupSafe==2.0.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
nose==1.3.7
notebook==6.4.11
numpy==1.22.3
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.0
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.20
Pygments==2.11.2
PyInputPlus==0.2.12
PyJWT==2.1.0
pyparsing==3.0.8
pyrsistent==0.18.1
PySimpleValidate==0.2.12
python-dateutil==2.8.2
python3-openid==3.2.0
python-dotenv==1.0.0
pytz==2021.1
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
requests==2.26.0
requests-oauthlib==1.3.0
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
Send2Trash==1.8.0
simplejson==3.18.0
six==1.16.0
sklearn==0.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.2.0
stdiomask==0.0.6
terminado==0.13.3
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.1.1
tornado==6.1
traitlets==5.1.1
uritemplate==3.0.1
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
python-dotenv==1.0.0
djoser==2.1.0
django-cors-headers==3.14.0
thefuzz[speedup]==0.19.0
